fix: move entity to script specific folder
