+++ src/domain/entities/user-monitoring/authorities-monitoring/UserRole.ts
