+++ src/domain/repositories/user-monitoring/authorities-monitoring/UserRolesRepository.ts
@@ -1,5 +1,5 @@
 import { Async } from "domain/entities/Async";
-import { UserRole } from "domain/entities/user-monitoring/UserRole";
+import { UserRole } from "domain/entities/user-monitoring/authorities-monitoring/UserRole";
 
 export interface UserRolesRepository {
     getByAuthorities(authority: string[]): Async<UserRole[]>;
